Initial version of the iOS data collection implementation
I haven't implemented the database layer, or the communication layer yet, so
the points are not stored or set to the server, but the data collection state
machine has been mostly implemented.

The one missing part is the trip end. Since we don't store points, we cannot
detect whether the last 15 trips are within the same 100 meter radius, so we
cannot automatically detect the end of a trip. So right now, we need to
manually signal trip end by using a button in the UI.

Database layer next and communication layer last