Ask permissions + turn on coarse location tracking
The initial install of the app on my phone worked without asking for
permissions, but didn't work on Mogeng's phone. We now ask for permission for:
- location tracking
- displaying notification and badges

We also turn on coarse location tracking, which should cause the app to get
restarted if it is killed or if the phone is rebooted.

Bonus fixes:
- On android, move elapsedRealtimeNanos into null check code to avoid crash
- On iOS, switch the development environment to 7.1
- Notify the user if activity tracking is unavailable