Get the google sign in to actually work
- Check in the .xib file for the google auth view. This was the real reason
  that the auth didn't work earlier.
- Change the main storyboard to wrap the view controller in a nav controller
  (Thanks to @neerajbaid for the tip on how to do this)
- Add the client ID and client secret for the google sign in
    - This necessitates pulling in the ConnectionSettings code as well
- Change the CommunicationHelper to take an array of trips to store, since the
  helper code associates the array with the "sections" key to make a JSON dict
- Replace the // Push trips here stub with the actual push

In addition, write the first serious location based tests, similar to the
current android tests. However, since iOS doesn't support mocking the location
interface, we do so in two steps:
- manually trigger state notifications using the notifiation handler
- manually invoke the delegate methods to mock location updates

In order to make this easier, expose some of the internal trip diary state
machine variables, namely the location manager and the activity manager as
@properties.

Fix a couple of issues found by the tests so far, mainly around the handling of
starting the tracking while currently in motion.

The current tests all pass. The ongoing trip test does not currently pass - the
corresponding code is commented out for checkin.

Next steps:
- Add tests for data untils, get the ongoing trip tests to work and fill in the
  endTrip test.