Turn on phone-based activity tracking on iOS
This commit turns on phone-based activity tracking on iOS.

However, since the CMMotionActivity class has only read-only properties, it was
fairly challenging to unit test it. I did so by creating a class of my own
(EMActivity, for E-Mission activity) that has very similar data. I also put the
data from the CMMotionActivityManager into the database and segment based on
the values in the database.

This allowed me to write a segmentation based unit test, which works.

When I tested it on a real iphone, however, I did not get any activity data for
the past 4 hours in response to my query. This was in spite of the iOS
documentation that said that the phone continuously collects activity data in
the background, even when not requested by any apps.

This needs to be investigated further, but I had to return my borrowed iPhone6.
Fortunately, test phones are "almost here".

In the meanwhile, if there is no activity information, we fall back on sending
the full set of GPS points for the trip to the server, since we should be able
to segment there using Francois' code.