Simplify end to end development by skipping authorization
When we are connected to a development server, skip the creation of the JWT and
just send the unencrypted email address instead. This makes it much easier to
start end to end development. Unfortunately, for iOS data collection, we still
need keys from parse, but that should not block most development.