Fix the long-term logging
We had long-term logging turned on for a while, trying to log to a file.
However, this didn't actually work, because the logger was reinitialized every
time the cfc_tracker process as initialized, but the handler was not
initialized because the preference was set.

However, even after removing the shared preferences check, this didn't work
because we would need to init the handler every time the logger was
initialized, which means every time the process was launched.

In order to avoid identifying all the potential entry points into the app, we
simply initialize the logger lazily. In order to do this along with the
FileHandler set, we need to change the log() method signatures to take a context.

This commit makes that change, and also provides convenience functions to use
the logger and to return log statements for display in the app.