Make the log files accessible via adb to help debugging
Before this, we stored the long term log files in the application internal
storage. Unfortunately, this is not accessible on physical devices unless the
device is rooted.

So we switch to the application external storage. This is less secure because
any application can read these files, and any application which can write to
external storage can write to it. But right now, debuggability trumps security,
at least while we are debugging the logging :)

Also refactored the log base to make it easier to store log files at various locations