Switch the user cache to use gson + implement methods to sync with the server
Switch the user cache to work directly with gson objects instead of json.
Also implement the outline of methods that need to be called by the data sync
code to sync to the server.

Note that none of this has been integrated or tested yet - that is the next set of
incoming changes.