Finish implementing and testing the sync to/from server methods
We now implement and test the sync to and from the server.
The main change between this and the previous change is that we no longer use
gson and the Entry class to serialize/deserialize the entries - instead we use
a manually generated json object.

This allows us to serialize/deserialize raw JSON strings, instead of requoting
them as "strings". In other words, we can load/store "{'metadata': {}, 'data':
{}}", not
{"data":"{\"mProvider\":\"TEST\",\"mResults\":[0.0,0.0],\"mAccuracy\":5.5....