Switch the existing data storage also to the new model
Add a couple of new methods to the user cache to support detecting the end of a
trip.  Largely remove all smoothing from the activity change code so that we
can get the raw, unadulterated feed, although we will probably want to
reintroduce some smoothing on the client at some point to avoid activity spew
on the server.