Add in support for a null context
When the TravelDiaryStateMachine is initialized using the no arg constructor,
as it does when the phone is restarted, the context is null. Ensure that we can
handle this case gracefully.

Also, add try/catch blocks around the log statements, so that an exception
while logging fails gracefully and does not crash the application.