Start adding consistency checks to the periodic invocation
While most issues encountered during the data collection have been fixed, there
are always going to be corner cases that we didn't see or that are not worth
handling (80:20 rule). For example, consider the case in which we create a
geofence after rebooting. Right now, we forcibly request a location update and
then create the geofence there. But what if we reboot the phone when we are
underground with no wifi? At that point, we may not get any location update,
and so maybe we won't be able to create the geofence.

In order to avoid ratholing into such issues, we instead periodically check
that the state machine is in a valid state and attempt to correct it if not.
This ensures that any inconsistency is only temporary.

For example, we check to see if the state machine is stuck in the start state,
and if so, attempt to initalize it - which means that our state machine will
eventually become valid.

TODO: We need to add more consistency checks here.