Fix native code crashes caused by serialization of standard android location class using gson
This fixes issue #49.

From repeated stacktraces, there as a pretty clear pattern of the crash
occuring when the location library interacted with the sqlite database. The
standard solution from the stackoverflow code, and the gson documentation,
would have been to write a custom serializer.

However, we don't invoke very many methods on the location classes - they are
really POJOs - so I just wrote a simple POJO wrapper that takes a standard
location as input and copies out the fields that we care about. This ensures
that we don't have to worry about the internal representation of the standard
location object.

The simple POJO wrapper is SimpleLocation. I've also changed places where
Location is written and read (Geofence creation and location handling) to use
the new class.

Testing done: if there is even one location that is NOT wrapped in
SimpleLocation (e.g. the one in the `GeofenceExitIntentService`, running the
tests in an emulator crashes within 15 mins. Once all of them have been
converted, the tests run for several hours.