Fixed bug + added timezone to usercache
Ensured that each usercache entry has a timezone in addition to the timestamp.
This allows us to localise the entries appropriately for display.

Also fix a fairly serious bug in the user cache - none of the sensed data was
being pushed! This is because the query to send the data to the server had not
been updated to include data of type "sensed_data". This is now fixed.