Allow customization of the location detection options
In order to explore the location detection space, change hardcoded values in
the location detection actions to be read from a config class instead. The
config class currently returns hardcoded values, but it would be fairly easy to
change it to read a value pushed from the server instead.

Most of the changes were pretty straightforward. The most significant change is
between the geofenced and ongoing trip detection. Since the complex state machine is
not required for ongoing collection, we create a new service
TripDiaryStateMachineOngoing which is invoked if we are not using geofencing.

We also need to stop checking for the trip end, which implies that we need to
remove the check for the last trip end transition while pushing data since it
will never be generated.