Remove a bunch of the old code dealing with detecting trip end and formatting the trips locally
This makes it so that the only code that is left
- deals with conversions to and from various data types, which seems to be
  relevant to "DataUtils"
- deals with pushing entries to the server

There's a little bit of trickiness in here though. Unlike on android, we cannot
get motion activity changes when the app is suspended. However, iOS
automatically tracks and retains this information for 7 days, so we can query
it while sending the data to the server.

There are some subtleties about querying this data at the end and sending it to the server.

First, how do we add it to the entries that are sent to the server? The
assumption on android was that all data would be in the database. We can
continue to maintain this invariant by storing the entries into the database
now, or we can wrap them with metadata and just add them to the data queried
from the database.

Here, we choose the second approach for two reasons:
- we can't use the generic "save" method anyway, since that would set all the
  write_ts of the metadata to the time that the data is read, and not to the
  time that the data was recorded.
- it seems like a waste in terms of performance to store a bunch of data into
  the database, turn around and read it immediately and then delete it after a
  successful send. Much better to avoid the overhead of the extra I/O, wrap it
  in a metadata object and append to the data from the database.

As usual, timezones are a problem. In particular, if we crossed timezones
during a trip, then the naive implementation would set the timezone for all
entries to be the timezone of the end of the trip, which would be wrong.

Fortunately, we do know what timezones we passed through during the trip - that
is encoded in the location data. So we use that to set the timezone for the
motion activities correctly.