Significantly cleanup the trip diary state machine init code
We were doing complicated stuff in the state machine init code depending on
whether this was a restart or not. But as we tested the code, the two branches
of the if converged, and the complexity made it harder to reason about the
code. So we radically simplify the code so that we can understand other stuff -
for example, why we get multiple calls for each state transition.