Ignore low quality points that cause the geofence exit
If the point that triggered the geofence exit has low accuracy, then we start
sensing again somewhat uselessly until we stop. Normally, this is quite quick,
but it is unnecessary, and can cause power drain if we are using high fidelity
data collection. And on android 6.0 in doze mode, it can take quite a bit of
time since medium accuracy data collection is automatically duty cycled.

So if we have a low accuracy exit, let us just skip it. The geofence still
exists, so we should continue to get exits, I think!?