Get the last n points based on the frequency of updates
Before, we used to get the last 10 points, because they represented the number
of points that we would expect to see in 5 mins, given a 30 sec interval. But
if we move to a 2 sec interval, we will see 10 points in less than minute, and
might detect the trip end too quickly.

So we change the number of points to check based on the frequency of updates.

We might also consider simply removing the number of points check and only
focus on the time based check, which seems more robust anyway.