Add a new helper class to determine the location at which to create a geofence
Before this, we used to create the geofence at the last location. But
sometimes, the last location is too old, and sometimes it is too inaccurate. We
really want a recent, accurate geofence - otherwise, we will not exit it
correctly and detect the start of a trip correctly.

If we want to read a new location for the geofence, we don't really need it to
match the criteria for the continuous data collection - for example, we may
have a large distance filter on the continuous data collection for lower power
drain, but that might prevent us from getting updates in time to create the new
geofence before the 30 sec deadline on the remote push is done.

So we create a helper utility, with its own location manager and delegate. This
location manager is configured for high accuracy and no filter, which improves
the chances that we will get a high accuracy point quickly. When a high
accuracy point is delivered to the delegate, it calls the geofence creation
code via an input callback.