Revert "Ignore low quality points that cause the geofence exit"
This reverts commit c57ef9d3b2387c7836e9c8b1571b5b4df2f689b7.

Revert change to ignore geofence exit for a low confidence point

This was a last minute change to try and reduce the time that we were in the
ongoing state. It is not clear that it actually helps or helped with power
drain.

The assumption behind the change was that if we didn't not respond to the
invalid geofence exit, then we would get a second exit if we were really
starting a trip.

However, this turned out to be false. Looking at the logs on my phone today, I saw this:

```
2:49pm: GeofenceExitIntentService: Geofence exited! triggering location accuracy = 897.0
2:49pm: Low accuracy point, Location[fused, 37.4041414, -122.083986 acc=897] ignoring
3:19pm: BuiltinUserCache: While searching for regex, got 0 results
3:19pm: BuiltinUserCache: While searching for all, got 0 results
3:19pm: BuiltinUserCache: There are no entries in the usercache. A sync must have just completed.
```

It looks like this is a premature optimization that should be removed ASAP