Use ms again while computing the number of points to return
In order to determine the number of points to return, we divide the time that
represents 5 minutes by the change detection interval. Recently, we changed the
time that represents 5 minutes to be in seconds instead of milliseconds,
consistent with our general change to store the data in seconds instead of
milliseconds, even on android.

But the change detection interval was still in milliseconds, since that is what
the API expects. This meant that we always retrieved 0 points, which meant that
we skipped the check.

So we convert the time in secs back to milliseconds before calculating the
number of points.