Fix the storage of the silentPushNotificationHandler
The biggest part was to store it correctly using the associated object.
I am not sure whether we need the get and set overrides - can't we just use self.XXX?
But when I tried to do that naively, I got errors about @dynamic, and this
works, so let's stick with it for now.