Ensure that we call the full background sync from the new serversync method for all background syncs
The primary exception is when a trip has just ended. At that point, we only
push the data, since we have already spent some time creating the geofence, and
we don't want to timeout of the 30 secs that we get.

For the background fetch (which is never executed) and the push notification
while waiting, we do a full background sync.

Tested using fake background syncs from the UI