Change the default location config to be consistent with android
= High accuracy, slow sampling

However, preliminary results from the data collection show that on iOS, the
sampling rate doesn't really affect the power drain - it is only the accuracy
that does. And if we don't have enough data points, iOS seems to miss some
trips.

So we'll increase the sampling frequency just a teeny little bit. And of
course, we need to investigate this further...