Fix compliation issues + issue with re-initializing on launch
Before this, we used the shared preferences to check whether initialization was
complete. The expectation was that the shared preferences would be set at the
end of the onboarding process
https://github.com/e-mission/e-mission-phone/blob/master/www/js/intro.js#L108
and we could just use it.

This preference seems to be set correctly, because when we read it here:
https://github.com/e-mission/e-mission-phone/blob/master/www/js/controllers.js#L10
it is true after the first init, but while reading it from native code, it is false.

This is probably because the app preferences appear to use local storage by
default.  Need to file an issue and investigate in depth. For now, we simply
set it from native code and move on.