Handle upgrades correctly
For a particular version of the app, once we have initialized the state
machine, it will run fine. But when we upgrade, all existing geofences, just
like all notifications, are deleted. But the shared preferences are NOT deleted.

So with the previous implementation, the FSM is initialized on install but not
on upgrade, which means that tracking stops on upgrade.

We fix that by storing the version that was initialized instead, and
reinitializing every time the version changes.