Displayed the length for the points not the distances
In 6f1f767a73c03a4ed360673d78aff59100ef2b5a, we tried to print the length of
the point arrays instead of the actual points. However, it turns out that we
printed the length of the _distance_ arrays and not the _point_ arrays.

Here, we go back to printing the whole distance array and print the length of
the point arrays instead.