Change the AppDelegate code to vastly simplify the startup logic
The current BEMAppDelegate is an implementation of some of the standard
AppDelegate methods for our app without modifying the original source code. It
does this by creating a category for the protocol.

https://developer.apple.com/library/ios/documentation/Cocoa/Conceptual/ProgrammingWithObjectiveC/CustomizingExistingClasses/CustomizingExistingClasses.html

However, while this allows us to provide our own implementation of methods that
are already defined in the protocol (e.g.
didRegisterForRemoteNotificationsWithDeviceToken), it does mean that we cannot
add any methods dynamically using a category, since a protocol is abstract,
unlike a class.

So in order to make the didFinishLaunchingWithOptions method callable from
pluginInitialize, we had to make it static. This was surprisingly easy because
there was very little in there that used delegate methods. In fact, we could
move the entire code out to another init-style class, although it is convenient
to have it here because everybody is used to looking for
`didFinishLaunchingWithOptions` in `AppDelegate.m`.

The one exception was the notification handler. The only reason that the
NotificationHandler needed to be an instance method instead of a static method
was that we needed to store the completion handler at the beginning and invoke
it after we have gone through several stages of the state machine.

In order to handle this, we moved the notification handling code to a separate
singleton class. The handler is now passed in through the userInfo, just like
it was when we tried to call it from the trip diary state machine. It is stored
while handling the first remote push, and invoked as usual at the end. This
means that we can process one remote push at a time, which seems like a fairly
reasonable assumption and one that we have always had.

Note that since we are passing the handler block around anyway, maybe we no
longer need to have two different notification handlers. Instead, we can have a
single handler that calls the block as needed. The challenge with that is to
ensure that we only call the handler block once.

With this structure, it is pretty clear that we want to call the handler block
only inside `handleNotifications`.