Read the location at which to create a geofence if needed
Fix the weird issue in which sometimes lastLocation is NULL and stays NULL for a while. This has been reported by:
- Hari, in whose case the lastLocation was NULL for over two weeks because he
  never uses google maps
- Andy, in whose case the lastLocation was NULL because he was underground and
  then I guess he never used google maps

This seems like a corner case, but since two people have encountered it, it
seems like a good idea to make it more robust.

Also, while I was in there anyway, also changed it so that we only created a
geofence at a valid point, and switched to this if the point was invalid. This
is less of an issue on android, because it turns out that we will get a
geofence exit immediately if we try to create a geofence at a point where we
are not actually present. For e.g. if we create a geofence in San Francisco
when we are in Mountain View, we will exit immediately and start tracking again.

But since we have similar code on iOS, I guess it's a good idea to add it here
as well, since we have to do all the work to read the location to handle the
NULL case anyway.