Revert "Run the initialization code in a separate thread"
This reverts commit a18f8f9385bdd9e37f7b412b386a911aee9a6ea0.

It turns out that if registered for from the background thread, geofences and
visit notifications are never triggered.

On a trip from home to Berkeley, with a 30 min stop at the education library in
the middle, neither the geofence, nor the visit notifications were triggered,
although visit notifications have been rock solid for us in the past.

On reverting the change, they were triggered again normally.