Fix the visit detection code so that it generates starts and ends correctly
Before this, we assumed that when a visit started, the departureDate was NULL,
since the messages on a visit started displayed a departureDate of '-'.
However, on additional logging, we found that this was actually NSDate distantFuture.

Fix the code to make the correct check, which now generates the correct transitions.

BONUS FIX:
- Add a geofence status check when we receive location updates in
  WAITING_FOR_TRIP_START.
- Remove obsolete method logPastHourCollectionCount which was generating a
  bunch of deprecated warnings, and was also unused