Add support for starting and ending tracking based on visit notifications
If it is enabled in the config, then turn tracking on and off based on visit
notifications alone. Also add support for the new DATA_PUSHED transition.

We need this because before this, we used to push in the remote push handlers
when the trip tracking ended. But now, in order to unify with the visit based
tracking, we push in the trip diary state machine handler. But if the
transition is launched from a remote sync, we still need to know when the
(async) push is complete so that we can set the callbackHandler appropriately.

So we create a new DATA_PUSHED transition.