Mechanical changes to match the new interface for the location config
Bunch of fairly mechanical changes to support the changes in
e7efe8eb0153308a5a7e0f40577b9717475423e0

The only non-mechanical change is to the plugin interface where, instead of
converting to json manually, we can just read directly from the cache, and we
can also now write to the cache.

This write is preferred to writing directly to the database, since it also
refreshes the cached copy of the config. Notably, we don't refresh the cached
copy every time at this point.

TODO: Figure out whether we should.