Add support for a new "simulateUser" option
Right now, when we want to simulate user access, we generate a local
notification including vibrate. We need to see whether this actually gets the
phone out of doze mode.

Since we currently only look at continuous tracking, we do not turn of the user
simulation at "night". But later, we may want to either represent the schedule
directly, or push it down from the server.