Add support for turning location tracking on and off
We already had support for turning tracking off, but then we just went back to
the start state. That is not a good solution because in order to make the
normal case more robust, we re-send the initialize transition if we are in the
start state.

So now, when we turn tracking off, we go to a different state
`tracking_stopped`, and the only way to get out of it is to recieve a new
`start_tracking` transition.

Also:
- refactored the plugin interface code to force general transitions instead of
  expanding the existing hardcoded list.
- ensured that the state machine is reset when the config changes through UI
  input by turning tracking on and off

The second of these is a workaround, not a real solution, since it doesn't deal
with automatic updates from the server that could come in as part of the sync,
and it is not clear that stopping and restarting tracking (or at least, doing
so in the current, hacky way) is the best way to reset.

Long-term solution is tracked at:
https://github.com/e-mission/e-mission-data-collection/issues/105