Stop restarting collection on reboot
This was a subtle bug. When the phone restarts, we don't know what state it
was in, but all the tracking that it was doing was turned off. So we turn on
tracking by sending an initialize from here, and handling the initialize code
specially in the services and having it always turn on the appropriate tracking
depending on the mode.

The problem with this that if the user has manually turned tracking off, then
we don't want this behavior. We handle this use case by checking for the
current state and only sending the initialize if the tracking has not stopped.