Turn off UI alerts unless the user wants them for debugging
User wants them == "simulateUserInteraction"
Unlike iOS, this does not introduce a new dependency, it does have a lot of
repeated code, specially in the services when we try to transition to a new
state.

Fixing it is being tracked in
https://github.com/e-mission/e-mission-data-collection/issues/113

Note that the iOS change changes the logging directly, so it is not checked
into this repo. The iOS change is at:
https://github.com/shankari/cordova-unified-logger/commit/f55fc2dd15f157b1d4a77c496d55a56f717621e2

This is an object case in the pitfalls of both approaches - we just need to
refactor and do them right on the second pass.