Old native code that is no longer used
These were files from the original native code that I wasn't sure that I had
ported over yet. But now we've been using the hybrid code for a while, so if
there's anything left behind, it is obsolete and can be removed.

And having this around is confusing - is this a plugin or a project? Are we
including dependencies from here?  ...

So let's remove them for now. We need to restore the tests, but that can wait a
bit.