Fix the crash on startup in iOS 8
I don't know how many more changes for iOS 8 are feasible, but this is an easy one so let's do it.
allowsBackgroundLocationUpdates is required in iOS9+ but is not supported in
iOS8, so we need to check whether it exists before calling it.

How far back do we need to go? pausesLocationUpdatesAutomatically is only
available from iOS6+, so we can't go further back than that. We don't really
support anything before iPhone5s, and those should all have come pre-installed
with at least iOS7,
https://www.amazon.com/dp/B00I2OOY26/ref=psdc_2407749011_t2_B00F3J4B5S
so I think that we are good.