Use the last 5 mins distances only when they actually represent 5 mins
We found this while investigating server side code. We check to see if the last
10 points and the points for the past 5 mins are all within the distance
filter. However, if we have navigation turned on, for example, we get points at
much shorter intervals, so we may have 10 points in less than 5 mins and end a
trip prematurely.

We fixed this on the server with
https://github.com/e-mission/e-mission-server/commit/231824b322ba7cbf792f3bf11feb78efec1faa7f

This is the client side fix, so that hopefully we will no longer have client ->
server mismatches caused by this.

This is the phone fix for https://github.com/e-mission/e-mission-server/issues/348