Ensure that we prompt for consent from remote push/visit notifications as well
We don't always load the full app if the app is in the background.
In practice, leaving the prompt only in the plugin caused it to never trigger,
as reported by users who did have the consent magically deleted.

Putting the checks in here caused us to have alerts when we
manually deleted the consent and roamed around.

TODO: We might want to consider moving the current consent requirements to an
xml/plist file instead of the current config.xml. That ensures that we can read
it in background native code without having to do creative stuff with XML parsers.