Add a description of why we use the motion sensors
This is to avoid the following error message from the play store.

> Missing Info.plist key - This app attempts to access privacy-sensitive data
> without a usage description. The app's Info.plist must contain an
> NSMotionUsageDescription key with a string value explaining to the user how the
> app uses this data.

The interesting thing is that this doesn't happen all the time, only
occasionally. But when it does happen, it is annoying because then I have to
bump up the version number and then the android and iOS version numbers are
slightly different occasionally.