Fix the first of many issues with location services tracking
This fixes the issue with turning location services off in the `WAITING_FOR_TRIP_START` state.
That generates a tracking error, which then generates a notification asking people to turn on location services again.
Assuming the checkForLocationSettings fails, the generated notification prompts the user to turn on location services again,
and the data collection plugin then re-initializes the FSM so that we return to `WAITING_FOR_TRIP_START`