Handle turning off location in the `ongoing_trip` state as well
The main difference is that the `checkLocationSettings` method now generates
the `tracking_error` transition, which basically just sets the state back to `start`.
We also perform actions on the `tracking_error`, primarily turning off any
existing tracking, to be consistent with being in the start state. This
involves passing in the final state to the action methods (e.g.
`deleteGeofence` and `stopAll`)

This inverts the logic - we generate the `tracking_error` from the check and
the handling of the tracking error involves turning off existing tracking and
moving to the start state.

This is a much better structure wrt decoupling. It also implies that when there
is an error with creating the geofence (e.g. in `createGeofenceInThread`, we
check the location settings instead of generating tracking error directly.

We also comment out the code that generates notifications while handling
`tracking_error`. Note that because of the inversion, we first generate the
resolution notificaton and **then** generate the `tracking_error`. This means
that if we generate a notification while processing the tracking error, it will
override the notification with the pending intent. We can fix this better once
we move to API 23+.
https://github.com/e-mission/e-mission-data-collection/issues/127#issuecomment-465383125

+ more logging around geofence creation, mainly checking for valid locations