Handle foregrounding of all services on O+
In O+, for apps compiled with api 26, a background app cannot launch a
background service. Our entire FSM runs in the background, so:
- all services need to foreground themselves on launch
- all services need to be launched as foreground services, but only on O+,
  because previous versions do not support start/get foreground service

This PR refactors the foreground service that runs when the trip is being
tracked, and creates static methods that then everybody else calls.

Please note that the diffs for `TripDiaryStateMachineForegroundService.java`
are a bit messy; `handleDestroy` is essentially a refactoring of `onDestroy`
into a static function and `getProperPendingIntent` is the new code.